requirements.txt
pandas==2.3.1
duckdb==1.3.2
openpyxl==3.1.5
python-dotenv==1.0.0
pyyaml==6.0.2